+++ packages/core/cli/src/__template__/src/components/Template.cy.ts
@@ -1,7 +1,6 @@
 // Cypress component test spec file
 
 import {%%COMPONENT_NAME%%} from './{%%COMPONENT_NAME%%}.vue'
-import { mount } from 'cypress/vue'
 
 describe('<{%%COMPONENT_NAME%%} />', () => {
   it('TODO: This is an example test', () => {